resticting access with permissions added

+++ src/blog/api/views.py
@@ -1,6 +1,7 @@
 from rest_framework import status
 from rest_framework.response import Response 
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
 
 from account.models import Account
 from blog.models import BlogPost
@@ -8,6 +9,7 @@ from blog.api.serializers import BlogPostSerializer
 
 
 @api_view(['GET', ])
+@permission_classes((IsAuthenticated, ))
 def api_detail_blog_view(request, slug):
 
     try: 
@@ -23,6 +25,7 @@ def api_detail_blog_view(request, slug):
 
 
 @api_view(['PUT', ])
+@permission_classes((IsAuthenticated, ))
 def api_update_blog_view(request, slug):
 
     try: 
@@ -31,6 +34,9 @@ def api_update_blog_view(request, slug):
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response': "You don't have permission to edit that. "})
 
     if request.method == 'PUT':
         serializer = BlogPostSerializer(blog_post, data=request.data)
@@ -40,11 +46,12 @@ def api_update_blog_view(request, slug):
             serializer.save()
             data['success'] = "update successful"
             return Response(data=data)
-        return Response(serializer.error, status=status.HTTP_400_BAD_REQUEST)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 
 @api_view(['DELETE', ])
+@permission_classes((IsAuthenticated, ))
 def api_delete_blog_view(request, slug):
 
     try: 
@@ -53,6 +60,9 @@ def api_delete_blog_view(request, slug):
     except BlogPost.DoesNotExist:
         return Response(status=status.HTTP_404_NOT_FOUND)
 
+    user = request.user
+    if blog_post.author != user:
+        return Response({'response': "You don't have permission to delete that. "})
 
     if request.method == 'DELETE':
         operation = blog_post.delete()
@@ -65,8 +75,10 @@ def api_delete_blog_view(request, slug):
 
 
 @api_view(['POST', ])
+@permission_classes((IsAuthenticated, ))
 def api_create_blog_view(request):
-    account = Account.objects.get(pk=1)
+    account = request.user
+    
     blog_post = BlogPost(author=account)
 
     if request.method == "POST":
